HikvisionTool/requirements.txt
re
request
tkinter
tkinterweb
threading
ipaddress
ipwhois
asyncio
aiohttp
itertools
Crypto.Cipher
cv2